refactor: separate CLI and core logic for better modularity
